+++ lending/src/sdk/models/shared/datastatus.ts
@@ -24,6 +24,7 @@ export enum DataTypes {
     DirectCosts = "directCosts",
     DirectIncomes = "directIncomes",
     Invoices = "invoices",
+    ItemReceipts = "itemReceipts",
     Items = "items",
     JournalEntries = "journalEntries",
     Journals = "journals",